Moved interactive pplanes to new file, added interactive pplane for ideal gas Euler
